Update floAuction.java
Commit 2/2.
Code reflects changes in language.yml.
Introduces %i %I and %e.
Alters how the code processes the displayName, giving it certain formats depending on whether it has a custom name or not.
Adds enchantIndicator to the message.

Strips messages of repeated whitespace; this was necessary as to avoid having 4 different alternative lines in language.yml. Without stripping, having %i %I %e would result in added whitespace between sections. This is also the reason why the formats had to have their own section in language.yml, because even if there wasn't a custom name and the displayName variable was "" empty, it still put the colour codes from the user's line definitions, making it impossible to strip the way I had thought of.